integrate topUpUSDT into deploy script
